correciones en las migraciones

+++ database/migrations/2022_09_16_045839_create_price_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('tourist_place_id');
             $table->string('description');
-            $table->integer('value');
+            $table->double('value', 8, 2);
             $table->timestamps();
         });
     }
